requirements.txt
asgiref==3.5.2
Django==4.1.1
djangorestframework==3.14.0
importlib-metadata==4.12.0
Markdown==3.4.1
pytz==2022.2.1
sqlparse==0.4.3
tzdata==2022.2
zipp==3.8.1
django-cors-headers==3.13.0
django-filter==22.1
gunicorn==20.1.0
urllib3==1.26.14
drf-yasg==1.21.5
graphene==3.2.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
